requirements.txt
# Core Dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
resampy>=0.4.2

# Signal Processing and Metrics
pypesq>=1.2.4
pystoi>=0.3.3
pesq>=0.0.4

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Deep Learning Utilities
einops>=0.6.1
timm>=0.9.0

# Configuration and Logging
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Progress Bars and CLI
tqdm>=4.65.0
click>=8.1.0

# Notebook Support
jupyter>=1.0.0
ipywidgets>=8.0.0
ipython>=8.12.0

# Utilities
pandas>=2.0.0
scikit-learn>=1.2.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Testing (optional)
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality (optional)
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0
